+++ CHANGELOG.md
@@ -4,11 +4,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [4.0.0] - 2021-04-08
 
--  The framework was updated to be compatible with Enact 4.0.0
+- The framework was updated to be compatible with Enact 4.0.0
 
 ## [4.0.0-alpha.1] - 2021-02-25
 
--  The framework was updated to support React 17.0.1
+- The framework was updated to support React 17.0.1
 
 ### Added